backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async Task Queue
celery[redis]==5.3.6
redis==5.0.1

# API Clients
requests==2.31.0
websocket-client==1.7.0
openai==1.10.0
aiohttp==3.9.1

# Audio/Video Processing
ffmpeg-python==0.2.0
pydub==0.25.1
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80

# Transcription & NLP
openai-whisper==20231117
whisperx==3.1.1
torch==2.1.2
torchaudio==2.1.2
transformers==4.37.2
sentencepiece==0.1.99

# Speaker Diarization
pyannote.audio==3.1.1
pyannote.core==5.0.0

# Emotion Detection
speechbrain==0.5.16

# Translation
sacremoses==0.1.1

# Face Detection (for Wav2Lip)
face-detection==0.2.2
face-recognition==1.3.0

# Deep Learning
tensorflow==2.15.0

# DeepFace (for gender detection)
deepface==0.0.86

# Utilities
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2
pillow==10.2.0
librosa==0.10.1

# Progress Tracking
tqdm==4.66.1

# Data Validation
python-multipart==0.0.6

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
